refactor: flat folder structure
