+++ src/utils.ts
@@ -1,6 +1,6 @@
 import assert from 'node:assert';
-import type { FinalOptions, Options } from '../types.js';
 import codecs from './codecs.js';
+import type { FinalOptions, Options } from './types.js';
 
 export const withDefaultOptions = (opt: Options): FinalOptions => {
   const options = typeof opt === 'string' ? { use: opt } : opt;
